form progress navigation
tracks the furthest step touched with form state. Need to pass a method that returns boolean if step should be green or not based on form validity and if it's a page that has been visited. Methods are set up, but wasn't able to pass to template using dynamic props.